[FE] fix: 비어있는 guildId로 요청 보내지는 문제 수정 (#69)
