+++ src/frontend/src/components/guild/GuildCategoriesList/index.tsx
@@ -2,10 +2,11 @@ import { BiHash } from 'react-icons/bi';
 import { BsFillMicFill } from 'react-icons/bs';
 import { TbPlus } from 'react-icons/tb';
 
+import { GuildChannelInfo, useChannelInfoStore } from '@/stores/channelInfo';
 import { useGuildInfoStore } from '@/stores/guildInfo';
 import useModalStore from '@/stores/modalStore';
 import { BodyMediumText, BodyRegularText } from '@/styles/Typography';
-import { CategoryDataResult, ChannelResult } from '@/types/guilds';
+import { CategoryDataResult, ChannelResult, ChannelType } from '@/types/guilds';
 
 import CreateChannelModal from '../CreateChannelModal';
 
@@ -15,11 +16,17 @@ export interface CategoriesListProps {
   categories?: CategoryDataResult[];
   channels?: ChannelResult[];
 }
-const CategoriesList = ({ categories, channels }: CategoriesListProps) => {
+const GuildCategoriesList = ({ categories, channels }: CategoriesListProps) => {
   const { openModal } = useModalStore();
   const { guildId } = useGuildInfoStore();
+  const { setSelectedChannel } = useChannelInfoStore();
+
   const handleOpenModal = (categoryId: string, guildId: string) => {
     openModal('withFooter', <CreateChannelModal categoryId={categoryId} guildId={guildId} />);
+  };
+
+  const handleChannelClick = (channelInfo: GuildChannelInfo) => {
+    setSelectedChannel({ id: channelInfo.id, name: channelInfo.name, type: channelInfo.type });
   };
 
   return (
@@ -34,7 +41,16 @@ const CategoriesList = ({ categories, channels }: CategoriesListProps) => {
             {channels
               ?.filter((channel) => category.categoryId === channel.categoryId)
               .map((channel) => (
-                <S.ChannelName key={channel.channelId}>
+                <S.ChannelName
+                  key={channel.channelId}
+                  onClick={() =>
+                    handleChannelClick({
+                      id: channel.channelId,
+                      name: channel.name,
+                      type: channel.channelType as ChannelType,
+                    })
+                  }
+                >
                   {channel.channelType === 'TEXT' ? <BiHash size={18} /> : <BsFillMicFill size={18} />}
                   <BodyRegularText>{channel.name}</BodyRegularText>
                 </S.ChannelName>
@@ -46,4 +62,4 @@ const CategoriesList = ({ categories, channels }: CategoriesListProps) => {
   );
 };
 
-export default CategoriesList;
+export default GuildCategoriesList;
